requirements.txt
streamlit
doctr
numpy
Pillow
